requirements.txt
python-dotenv~=0.21.0
https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip
fluentogram
Babel